+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 6.7.8
 ### Documentation
 - Remove mentions of `kafka` 3.7.0 version as it is no longer supported.
+### Fixes
+- Fix S3 key test, add back update test
 
 ## 6.7.7
 ### Features