+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 6.7.8
+### Chore
+- Update `github.com/hashicorp/terraform-plugin-sdk/v2` dependency to latest version
 ### Documentation
 - Remove mentions of `kafka` 3.7.0 version as it is no longer supported.
+- Update import method for `ionoscloud_datacenter` and `ionoscloud_server` resources.
 
 ## 6.7.7
 ### Features